usuario e outras coisas mais

+++ src/main/java/br/com/bsitecnologia/dashboard/controller/BaseCrudBean.java
@@ -6,6 +6,7 @@ import java.util.List;
 
 import javax.faces.application.FacesMessage;
 import javax.faces.application.FacesMessage.Severity;
+import javax.faces.component.UIViewRoot;
 import javax.faces.context.FacesContext;
 import javax.faces.event.ValueChangeEvent;
 import javax.faces.model.SelectItem;
@@ -48,8 +49,15 @@ public abstract class BaseCrudBean<T extends BaseEntity> implements Serializable
 	
 	protected void postRowSelect(){}
 	protected void postLoad(){}
-	protected void postPersist(){}
 	protected void prePersist(){}
+	protected void postPersist(){}
+	protected void preUpdate(){}
+	protected void postUpdate(){}
+	protected void preDelete(){}
+	protected void postDelete(){}
+	protected boolean persistValidation(){return true;}
+	protected boolean updateValidation(){return true;}
+	protected boolean deleteValidation(){return true;}
 	
 	
 	/*Ações comuns a serem herdadas*/
@@ -65,25 +73,54 @@ public abstract class BaseCrudBean<T extends BaseEntity> implements Serializable
 	}
 	
 	public void save(){
-		prePersist();
-		getDao().save(getFormEntity());
-		loadList();
-		addMessage(FacesMessage.SEVERITY_INFO, String.format("%s: %s", title, getFormEntity().getEntityDescription()), String.format("%s salvo(a) com sucesso.", title));
-		saveButtonLabel = Buttons.SAVE.getLabel();
-		showDeleteButton = false;
-		postPersist();
-		resetFormEntity();
+		if(executeValidations()){
+			executePreInterceptors();
+			setFormEntity(getDao().save(getFormEntity()));
+			loadList();
+			addMessage(FacesMessage.SEVERITY_INFO, String.format("%s: %s", title, getFormEntity().getEntityDescription()), String.format("%s salvo(a) com sucesso.", title));
+			saveButtonLabel = Buttons.SAVE.getLabel();
+			showDeleteButton = false;
+			executePostInterceptors();
+			resetFormEntity();
+		}
+	}
+	
+	private void executePostInterceptors() {
+		if(showDeleteButton){
+			postUpdate();
+		}else{
+			postPersist();
+		}
+		
+	}
+	
+	private void executePreInterceptors() {
+		if(showDeleteButton){
+			preUpdate();
+		}else{
+			prePersist();
+		}
+	}
+	
+	private boolean executeValidations() {
+		if(showDeleteButton){
+			return updateValidation();
+		}else{
+			return persistValidation();
+		}
 	}
 	
 	public void delete(){
-		prePersist();
-		getDao().delete(getFormEntity());
-		loadList();
-		addMessage(FacesMessage.SEVERITY_INFO,  String.format("%s: %s", title, getFormEntity().getEntityDescription()), String.format("%s deletado(a) com sucesso.", title));
-		saveButtonLabel = Buttons.SAVE.getLabel();
-		showDeleteButton = false;
-		postPersist();
-		resetFormEntity();
+		if(deleteValidation()){
+			preDelete();
+			getDao().delete(getFormEntity());
+			loadList();
+			addMessage(FacesMessage.SEVERITY_INFO,  String.format("%s: %s", title, getFormEntity().getEntityDescription()), String.format("%s deletado(a) com sucesso.", title));
+			saveButtonLabel = Buttons.SAVE.getLabel();
+			showDeleteButton = false;
+			postDelete();
+			resetFormEntity();
+		}
 	}
 	
 	public void cancel(){
@@ -103,7 +140,6 @@ public abstract class BaseCrudBean<T extends BaseEntity> implements Serializable
 		FacesContext context = FacesContext.getCurrentInstance();
 		FacesMessage fm =  new FacesMessage(severity, summary, detail);
         context.addMessage(null, fm);
-        context.renderResponse();
 	}
 	
 	protected <A extends BaseEntity> A getEntityFromValueChangeEvent(ValueChangeEvent event, List<A> list){
@@ -117,12 +153,30 @@ public abstract class BaseCrudBean<T extends BaseEntity> implements Serializable
 		return null;
 	}
 	
+	protected <A extends BaseEntity> A getEntityById(Integer id, List<A> list){
+		for (A a : list) {
+			if(id.equals( a.getId() ) ){
+				return a;
+			}
+		}
+		return null;
+	}
+	
+	
 	protected List<SelectItem> fillSelectItemList(List<? extends BaseEntity> list) {
 		List<SelectItem> selectItemList = new ArrayList<SelectItem>();
 		for(BaseEntity t: list){
 			selectItemList.add(new SelectItem(t.getId(), t.getEntityDescription()));
 		}
 		return selectItemList;
+	}
+	
+	protected void redirect(String page){
+		//			redirect("/admin/usuario/usuario");
+		FacesContext facesContext = FacesContext.getCurrentInstance();
+		UIViewRoot viewRoot = facesContext.getApplication().getViewHandler().createView(facesContext, page);
+		facesContext.setViewRoot(viewRoot);
+		facesContext.renderResponse();
 	}
 	
 	/*get&set*/
